flyttet henting av aktivitet data inn i sine repositories
