+++ src/main/java/no/nav/pto/veilarbportefolje/postgres/AktoerDataOpensearchMapper.java
