+++ src/main/java/no/nav/pto/veilarbportefolje/postgres/utils/AktivitetStatusData.java
